Support subtables that precede the parent id column in XML order
This feature will defer processing of subtables that require the
parent table id column (one-to-many and many-to-many relations)
and which precede the id column in the XML document order.

This is done by buffering the subtable XML in memory and applying
it after the parent table has finished processing. As such this will
increase memory usage based on the maximum size of the subtable fields.